rag_pdf/requirements.txt
streamlit
python-dotenv
pypdf
 
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-pinecone
 
pinecone
sentence-transformers
 